refactor + requested changes
